refactor: centralize backend checks and toasts
