+++ login.js
@@ -20,6 +20,7 @@ import {
 import { firebaseConfig, setPassphrase, getPassphrase, clearPassphrase } from './firebase-config.js';
 import { encryptString, decryptString } from './crypto.js';
 import { fetchResponsavelFinanceiroUsuarios } from './responsavel-financeiro.js';
+import { showToast } from './utils.js';
 
 const app = getApps().length ? getApps()[0] : initializeApp(firebaseConfig);
 const auth = getAuth(app);
@@ -35,45 +36,6 @@ let selectedRole = null;
 window.isFinanceiroResponsavel = false;
 window.responsavelFinanceiro = null;
 
-// Ping a local file to test online status without CORS issues.
-async function pingOnline(timeout = 3000) {
-  try {
-    const ctrl = new AbortController();
-    const t = setTimeout(() => ctrl.abort(), timeout);
-    const res = await fetch('ping.txt?cb=' + Date.now(), {
-      cache: 'no-store',
-      signal: ctrl.signal
-    });
-    clearTimeout(t);
-    return res.ok;
-  } catch {
-    return false;
-  }
-}
-
-// Verifica conectividade com backend/API.
-export async function checkBackend() {
-  if (!navigator.onLine) return false;
-  return await pingOnline();
-}
-
-
-function showToast(message, type = 'success') {
-  const container = document.getElementById('toastContainer');
-  if (!container) {
-    alert(message);
-    return;
-  }
-  const color = type === 'success' ? 'bg-green-500'
-               : type === 'error' ? 'bg-red-500'
-               : type === 'warning' ? 'bg-yellow-500'
-               : 'bg-blue-500';
-  const toast = document.createElement('div');
-  toast.className = `${color} text-white px-4 py-3 rounded shadow-lg flex items-center justify-between min-w-[250px] animate-fadeIn`;
-  toast.innerHTML = `<span class="mr-4">${message}</span><button onclick="this.parentElement.remove()" class="text-white font-bold">×</button>`;
-  container.appendChild(toast);
-  setTimeout(() => toast.remove(), 4000);
-}
 
 window.savePassphrase = () => {
   const input = document.getElementById('passphraseInput');
