+++ utils.js
@@ -1,0 +1,40 @@
+export async function pingOnline(timeout = 3000) {
+  try {
+    const ctrl = new AbortController();
+    const t = setTimeout(() => ctrl.abort(), timeout);
+    const res = await fetch('ping.txt?cb=' + Date.now(), {
+      cache: 'no-store',
+      signal: ctrl.signal
+    });
+    clearTimeout(t);
+    return res.ok;
+  } catch {
+    return false;
+  }
+}
+
+export async function checkBackend() {
+  if (!navigator.onLine) return false;
+  return await pingOnline();
+}
+
+export function showToast(message, type = 'success') {
+  const container = document.getElementById('toastContainer');
+  if (!container) {
+    alert(message);
+    return;
+  }
+  const color =
+    type === 'success'
+      ? 'bg-green-500'
+      : type === 'error'
+      ? 'bg-red-500'
+      : type === 'warning'
+      ? 'bg-yellow-500'
+      : 'bg-blue-500';
+  const toast = document.createElement('div');
+  toast.className = `${color} text-white px-4 py-3 rounded shadow-lg flex items-center justify-between min-w-[250px] animate-fadeIn`;
+  toast.innerHTML = `<span class="mr-4">${message}</span><button onclick="this.parentElement.remove()" class="text-white font-bold">×</button>`;
+  container.appendChild(toast);
+  setTimeout(() => toast.remove(), 4000);
+}
